fix: 아키텍쳐 구조 맞춰서 변경
코멘트 기준으로 변경 했습니다.
